+++ src/domain/raum/index.ts
@@ -1,0 +1,3 @@
+export * from './raum';
+export * from './raum-repository-port';
+export * from './raum-service';